Add files via upload
In this supervised machine learning model, the goal is to predict a student's score based on the number of hours they spend studying. The model is trained using a dataset that includes historical data, where each entry consists of a pair of values: the number of hours a student studied and the corresponding score they achieved on an exam.